refactor: Improve UI and station display
Refactor MainActivity to use Compose and MVVM architecture. Adjust Play/Pause button to take up 50% of screen height and optimize station departure display to show only next departure per line. Add proper delay indicators and improve loading states. Fix location permission handling and station distance calculation.
